move page loader out of folder
