+++ client/src/App.tsx
@@ -8,7 +8,7 @@ import StatisticsPage from "./pages/Statistics/StatisticsPage";
 import FriendsPage from "./pages/Friends/FriendsPage";
 import SettingsPage from "./pages/Settings/SettingsPage";
 import { useAuth0 } from "@auth0/auth0-react";
-import { PageLoader } from "./components/PageLoader/PageLoader";
+import { PageLoader } from "./components/PageLoader";
 import { AuthenticationGuard } from "./utils/Auth0/AuthenticationGuard";
 import { CallbackPage } from "./non-app-pages/Callback/CallbackPage";
 import { NotFoundPage } from "./non-app-pages/NotFound/NotFoundPage";
